+++ x/lightclient/keeper/ibc_msg_channel_open_ack.go
@@ -1,4 +1,4 @@
-package ante
+package keeper
 
 import (
 	errorsmod "cosmossdk.io/errors"
