+++ x/lightclient/keeper/ibc_msg_update_client_test.go
@@ -1,4 +1,4 @@
-package ante_test
+package keeper_test
 
 import (
 	"testing"
@@ -13,7 +13,7 @@ import (
 	ibcsolomachine "github.com/cosmos/ibc-go/v8/modules/light-clients/06-solomachine"
 	ibctm "github.com/cosmos/ibc-go/v8/modules/light-clients/07-tendermint"
 	keepertest "github.com/dymensionxyz/dymension/v3/testutil/keeper"
-	"github.com/dymensionxyz/dymension/v3/x/lightclient/ante"
+	"github.com/dymensionxyz/dymension/v3/x/lightclient/keeper"
 	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
 	"github.com/stretchr/testify/require"
 )
@@ -122,7 +122,7 @@ func TestHandleMsgUpdateClientGood(t *testing.T) {
 	ibcclientKeeper := NewMockIBCClientKeeper(testClientStates)
 	ibcchannelKeeper := NewMockIBCChannelKeeper(nil)
 	rollappKeeper := NewMockRollappKeeper(rollapps, stateInfos)
-	ibcMsgDecorator := ante.NewIBCMessagesDecorator(*k, ibcclientKeeper, ibcchannelKeeper, rollappKeeper)
+	ibcMsgDecorator := keeper.NewIBCMessagesDecorator(*k, ibcclientKeeper, ibcchannelKeeper, rollappKeeper)
 	clientMsg, err := ibcclienttypes.PackClientMessage(&header)
 	require.NoError(t, err)
 	msg := &ibcclienttypes.MsgUpdateClient{
