+++ x/lightclient/keeper/ibc_msgs_test.go
@@ -1,4 +1,4 @@
-package ante_test
+package keeper_test
 
 import (
 	"context"
